FBO can work with texture id, if we don't have Texture.
It will help with Oculus 0.6.0.0 integration, because we get only texture ids and we need to render to them.

Other (more flexible) option would be to convert texture id to Texture. But it would bring some allocation complications/overhead, so I decided for working with raw texture ids, at least for now, as it requires only tiny header change.