Fix ubershader illumination showing up on faces oriented away from light.
svn5624 in 2012-06-04 would be correct only if pow(negative,) preserves sign. And maybe it did, back then. But it is undefined, and it does not preserve sign now. So ilumination leaks little bit (shadow bias) behind corners. Leak was visible in Lightsmark frame 23 (and probably in many scenes in SV/RL).